requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-graphql-jwt==0.4.0
exceptiongroup==1.1.3
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
iniconfig==2.0.0
packaging==23.1
pkg_resources==0.0.0
pluggy==1.3.0
promise==2.3
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
